# Change release builds to be optimized