--- a/src/NUnitFramework/testdata.fsharp/nunit.testdata.fsharp.fsproj
+++ b/src/NUnitFramework/testdata.fsharp/nunit.testdata.fsharp.fsproj
@@ -6,6 +6,7 @@
     <DisableImplicitFSharpCoreReference>true</DisableImplicitFSharpCoreReference>
     <RestoreIgnoreFailedSources>true</RestoreIgnoreFailedSources>
     <NoWarn>$(NoWarn);NU1604</NoWarn>
+    <Optimize>false</Optimize>
   </PropertyGroup>
 
   <ItemGroup>