As discussed in #4149 , seems like a good time with 4.0 nearing to try out again distributing optimized release builds. With this PR removing the forced no-optimize and BenchmarkDotNet project proves that artifacts can now be consumed as release builds (no longer need to disable validator).

Quick benchmark test for changes. Shows > 10% time reduction in best cases (very feature-limited benchmarks of course).

``` ini

BenchmarkDotNet=v0.13.5, OS=Windows 11 (10.0.22621.1555/22H2/2022Update/SunValley2)
12th Gen Intel Core i9-12900H, 1 CPU, 20 logical and 14 physical cores
.NET SDK=7.0.203
  [Host] : .NET 6.0.16 (6.0.1623.17311), X64 RyuJIT AVX2

Job=InProcess  Toolchain=InProcessEmitToolchain  

```

## Before (master)

|                       Method |        Mean |       Error |      StdDev |      Gen0 |     Gen1 |     Gen2 |  Allocated |
|----------------------------- |------------:|------------:|------------:|----------:|---------:|---------:|-----------:|
|        CountTestsEmptyFilter |    138.5 us |     1.89 us |     1.77 us |    4.1504 |        - |        - |    51.6 KB |
| CountTestsWithCategoryFilter |  2,758.3 us |    52.83 us |    49.41 us |  500.0000 |        - |        - | 6172.07 KB |
|    CountTestsWithOrFilterAll |  4,806.7 us |    58.17 us |    51.56 us |  312.5000 | 125.0000 |  23.4375 | 3903.07 KB |
|                    LoadTests | 66,056.0 us | 1,224.24 us | 1,202.37 us | 2125.0000 | 875.0000 | 250.0000 | 24510.9 KB |

## After (this PR)

|                       Method |        Mean |       Error |    StdDev |      Gen0 |      Gen1 |     Gen2 |   Allocated |
|----------------------------- |------------:|------------:|----------:|----------:|----------:|---------:|------------:|
|        CountTestsEmptyFilter |    116.8 us |     1.72 us |   1.61 us |    4.1504 |         - |        - |     51.6 KB |
| CountTestsWithCategoryFilter |  2,332.5 us |    28.56 us |  26.72 us |  500.0000 |         - |        - |  6172.07 KB |
|    CountTestsWithOrFilterAll |  3,921.2 us |    76.52 us |  71.58 us |  312.5000 |  125.0000 |  23.4375 |  3903.08 KB |
|                    LoadTests | 61,599.0 us | 1,057.11 us | 988.82 us | 2222.2222 | 1000.0000 | 333.3333 | 24131.66 KB |