@lahma 2023-04-14
So test projects can also be optimized. Had to fix two things:

* `SlowPath_TakenWhenSpecialTypes` verifies code path selection via stopwatch (lovely), so lowered the value from 50ms to 30ms as optimized code is fast(!)
* `DefaultConstraintTests.Success_DefaultStructWithOverriddenEquals()` has different `ToString` on full framework for some reason under optimized build but this is the behavior I believe NUnit consumer will see anyway because the type in question will be optimized based on their project configuration